+++ components/roast-record-list/RoastRecordCard.tsx
@@ -1,0 +1,112 @@
+'use client';
+
+import { formatTime } from '@/lib/roastTimerUtils';
+import { Card } from '@/components/ui';
+import { HiTrash, HiCalendar } from 'react-icons/hi';
+import { MdTimer } from 'react-icons/md';
+import { PiCoffeeBeanFill } from 'react-icons/pi';
+import type { RoastTimerRecord } from '@/types';
+
+interface RoastRecordCardProps {
+  record: RoastTimerRecord;
+  onDelete: (id: string, e: React.MouseEvent) => void;
+  onClick: (id: string) => void;
+}
+
+const formatDate = (dateStr: string) => {
+  // YYYY-MM-DD形式をYYYY/MM/DD形式に変換
+  const [year, month, day] = dateStr.split('-');
+  return `${year}/${month}/${day}`;
+};
+
+const getRoastLevelColor = (
+  level: '浅煎り' | '中煎り' | '中深煎り' | '深煎り'
+) => {
+  switch (level) {
+    case '深煎り':
+      return '#120C0A';
+    case '中深煎り':
+      return '#4E3526';
+    case '中煎り':
+      return '#745138';
+    case '浅煎り':
+      return '#C78F5D';
+    default:
+      return '#6B7280';
+  }
+};
+
+export function RoastRecordCard({ record, onDelete, onClick }: RoastRecordCardProps) {
+  return (
+    <Card
+      variant="hoverable"
+      className="p-3 md:p-4 relative h-auto"
+      onClick={() => onClick(record.id)}
+    >
+      {/* 削除ボタン（右上） */}
+      <button
+        onClick={(e) => onDelete(record.id, e)}
+        className="absolute top-2 right-2 p-1.5 text-red-600 hover:bg-red-50 rounded-lg transition-colors min-h-[32px] min-w-[32px] flex items-center justify-center z-10"
+        aria-label="削除"
+      >
+        <HiTrash className="h-4 w-4" />
+      </button>
+
+      {/* 豆名と焙煎度合い */}
+      <div className="flex items-center gap-2 mb-3 pr-8">
+        <div className="flex-shrink-0">
+          <PiCoffeeBeanFill className="h-4 w-4 md:h-5 md:w-5 text-amber-700" />
+        </div>
+        <div className="flex-1 min-w-0 flex items-center gap-2 flex-wrap">
+          <h3 className="text-base md:text-lg font-bold text-gray-900 truncate">
+            {record.beanName}
+          </h3>
+          <span
+            className="inline-block px-2 py-0.5 text-white text-xs font-semibold rounded-full flex-shrink-0"
+            style={{ backgroundColor: getRoastLevelColor(record.roastLevel) }}
+          >
+            {record.roastLevel}
+          </span>
+        </div>
+      </div>
+
+      {/* 詳細情報 */}
+      <div className="space-y-2">
+        {/* 焙煎時間 */}
+        <div className="flex items-center gap-2 text-gray-700">
+          <MdTimer className="h-4 w-4 text-gray-400 flex-shrink-0" />
+          <div className="flex items-center gap-2">
+            <span className="text-xs text-gray-500">焙煎時間</span>
+            <span className="text-sm md:text-base font-medium text-gray-900 font-mono">
+              {formatTime(record.duration)}
+            </span>
+          </div>
+        </div>
+
+        {/* 重さ */}
+        <div className="flex items-center gap-2 text-gray-700">
+          <div className="h-4 w-4 flex items-center justify-center flex-shrink-0">
+            <span className="text-gray-400 text-base">⚖</span>
+          </div>
+          <div className="flex items-center gap-2">
+            <span className="text-xs text-gray-500">重さ</span>
+            <span className="text-sm md:text-base font-medium text-gray-900 font-mono">
+              {record.weight}g
+            </span>
+          </div>
+        </div>
+
+        {/* 焙煎日 */}
+        <div className="flex items-center gap-2 text-gray-700">
+          <HiCalendar className="h-4 w-4 text-gray-400 flex-shrink-0" />
+          <div className="flex items-center gap-2">
+            <span className="text-xs text-gray-500">焙煎日</span>
+            <span className="text-sm md:text-base font-medium text-gray-900 font-mono">
+              {formatDate(record.roastDate)}
+            </span>
+          </div>
+        </div>
+      </div>
+    </Card>
+  );
+}
